requirements.txt
streamlit
tensorflow==2.12.0
Pillow
numpy
